move hash_address_to_partition to epoch_rewards_hasher.rs
